# Use `include?` and `find` for performance