Fix a couple of performance RuboCop offenses in workbook

```
Warming up --------------------------------------
    match_and_select    59.143k i/100ms
    include_and_find   123.780k i/100ms
Calculating -------------------------------------
    match_and_select    591.904k (± 1.0%) i/s -      3.016M in   5.096401s
    include_and_find      1.238M (± 0.8%) i/s -      6.313M in   5.098961s

Comparison:
    include_and_find:  1238131.2 i/s
    match_and_select:   591904.2 i/s - 2.09x  (± 0.00) slower

Calculating -------------------------------------
    match_and_select   947.000  memsize (   499.000  retained)
                         9.000  objects (     1.000  retained)
                         3.000  strings (     0.000  retained)
    include_and_find   160.000  memsize (     0.000  retained)
                         4.000  objects (     0.000  retained)
                         2.000  strings (     0.000  retained)

Comparison:
    include_and_find:        160 allocated
    match_and_select:        947 allocated - 5.92x more
```

### Your checklist for this pull request
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] I have reviewed the [guidelines for contributing](../blob/master/CONTRIBUTING.md) to this repository.
- [ ] I have added (or updated) appropriate tests if this PR fixes a bug or adds a feature.
- [ ] If this PR doesn't need tests (docs change), I added `[ci skip]` to the title of the PR.
- [ ] If this closes any issues, I have added "Closes `#issue`" to the PR description or my commit messages.
- [ ] I have updated the documentation accordingly.
- [x] All new and existing tests passed.
- [ ] I added an entry to the [changelog](../blob/master/CHANGELOG.md).